Bug fixes and additions
LibraryOrganizer will be a huge problem and need to continually be
checked. For now it’s again working and will clean itself of .DS_Store
files. Files besides m4a’s and .DS_Store will crash it. LibraryLoader
was added to instantiate a library object that holds all of the tracks
and unique artists, albums, ect. in the library. Songs will be taken
from the library to be played on wherever music ends up being played
from.