Lotsa Server Stuff + More
The server does things but nothing that actually impacts how it
operates as a server. Essentially right now the server can now operate
as a standalone application. It’s got a chatroom which would function
at least send to a working client (but not receive) and a consolelite
version of what I was previously working on. It will be nice to
hopefully eventually make it so that the server gets meaningful
messages back from the music player instead of just lines like “play”,
“stopped” and “pause”. Also I need to add a whole actually user
interface for that user interface stuff…

There’s a demo to get metadata from mp3 files but since I don’t really
have a library of mp3s at all, it might be a while until I’m actually
convinced to add them.

It’s been a while since a commit so there’s some other features too.
The console now performs find “<album>”… “<track>” functions and I
revamped the add function to work with the library object instead of
just looking all over the file system again. Let’s only mess with I/O
when we need to.

Library also has a few more useful functions that utilizes the data it
previously finds when it formats the library.

Next up is the client.