Client/Server Interactions
Setup the client with a GUI and started working on getting the two
parts to interact with each other. Currently the server and client can
send chat messages to each other and the client can send DDJConsole
commands to the server to control the music.

So I’ve got one of the major functionalities that I was looking to
actually make with this project - have a client control a server music
player. With this, I’ll be able to hook my stereo to a raspberry pi and
control what it plays from my computer. It works locally so I’m
assuming that it’ll work if I’m simply hosting across the room but
it’ll need some testing. I’m worried about -
line = AudioSystem.getSourceDataLine(aufmt); in the playback thread.
Does that just pull default system audio from wherever the program is
running from?

I want to rewrite a lot of the utilities from JAAD (Java Advanced Audio
Decoder library) further into this project, maybe even reformat mp3s
and m4as into some sort of other music file which would be used by my
program (the idea is to make the metadata easy to read, and the music
data readily accessible to be streamed locally or over a TCP
connection)